back-end/requirements.txt
appdirs==1.4.4
attrs==20.3.0
bandit==1.7.0
bidict==0.21.2
black==20.8b1
click==7.1.2
Flask==1.1.2
flask-marshmallow==0.14.0
Flask-SocketIO==5.0.1
Flask-SQLAlchemy==2.4.4
gitdb==4.0.5
GitPython==3.1.12
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
marshmallow==3.10.0
marshmallow-sqlalchemy==0.24.1
mypy-extensions==0.4.3
packaging==20.8
pathspec==0.8.1
pbr==5.5.1
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.1
python-dateutil==2.8.1
python-engineio==4.0.0
python-socketio==5.0.4
PyYAML==5.4.1
regex==2020.11.13
six==1.15.0
smmap==3.0.4
SQLAlchemy==1.3.22
stevedore==3.3.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
waitress==1.4.4
Werkzeug==1.0.1
